requirements.txt
# FastAPI stack
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
SQLAlchemy==2.0.25

# Pydantic ecosystem (v2)
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-settings==2.2.1

# Security / Auth
passlib[bcrypt]==1.7.4
bcrypt==4.1.2            # Explicitly pin bcrypt to avoid backend errors
PyJWT==2.8.0

# File uploads
python-multipart==0.0.9

# Testing
pytest==8.0.2
httpx==0.27.0
